Fix scaler crash
