+++ scaler/handlers.go
@@ -65,13 +65,15 @@ func (e *impl) IsActive(
 			Result: true,
 		}, nil
 	}
-	allCounts := mergeCountsWithRoutingTable(
+
+	hostCount, ok := getHostCount(
+		host,
 		e.pinger.counts(),
 		e.routingTable,
 	)
-	hostCount, ok := allCounts[host]
 	if !ok {
 		err := fmt.Errorf("host '%s' not found in counts", host)
+		allCounts := mergeCountsWithRoutingTable(e.pinger.counts(), e.routingTable)
 		lggr.Error(err, "Given host was not found in queue count map", "host", host, "allCounts", allCounts)
 		return nil, err
 	}
@@ -160,16 +162,18 @@ func (e *impl) GetMetrics(
 		lggr.Error(err, "ScaledObjectRef", metricRequest.ScaledObjectRef)
 		return nil, err
 	}
-	allCounts := mergeCountsWithRoutingTable(
+
+	hostCount, ok := getHostCount(
+		host,
 		e.pinger.counts(),
 		e.routingTable,
 	)
-	hostCount, ok := allCounts[host]
 	if !ok {
 		if host == "interceptor" {
 			hostCount = e.pinger.aggregate()
 		} else {
 			err := fmt.Errorf("host '%s' not found in counts", host)
+			allCounts := mergeCountsWithRoutingTable(e.pinger.counts(), e.routingTable)
 			lggr.Error(err, "allCounts", allCounts)
 			return nil, err
 		}
