+++ scaler/main.go
@@ -125,6 +125,7 @@ func main() {
 			cfg,
 			healthPort,
 			pinger,
+			table,
 		)
 	})
 	lggr.Error(grp.Wait(), "one or more of the servers failed")
@@ -172,6 +173,7 @@ func startAdminServer(
 	cfg *config,
 	port int,
 	pinger *queuePinger,
+	routingTable *routing.Table,
 ) error {
 	lggr = lggr.WithName("startHealthcheckServer")
 
@@ -184,7 +186,7 @@ func startAdminServer(
 	})
 	mux.HandleFunc("/queue", func(w http.ResponseWriter, r *http.Request) {
 		lggr = lggr.WithName("route.counts")
-		cts := pinger.counts()
+		cts := mergeCountsWithRoutingTable(pinger.counts(), routingTable)
 		lggr.Info("counts endpoint", "counts", cts)
 		if err := json.NewEncoder(w).Encode(&cts); err != nil {
 			lggr.Error(err, "writing counts information to client")
@@ -200,7 +202,7 @@ func startAdminServer(
 			w.Write([]byte("error requesting counts from interceptors"))
 			return
 		}
-		cts := pinger.counts()
+		cts := mergeCountsWithRoutingTable(pinger.counts(), routingTable)
 		lggr.Info("counts ping endpoint", "counts", cts)
 		if err := json.NewEncoder(w).Encode(&cts); err != nil {
 			lggr.Error(err, "writing counts data to caller")
